+++ kompendium-annotations/src/main/kotlin/io/bkbn/kompendium/annotations/FreeFormObject.kt
@@ -1,0 +1,5 @@
+package io.bkbn.kompendium.annotations
+
+@Retention(AnnotationRetention.RUNTIME)
+@Target(AnnotationTarget.PROPERTY)
+annotation class FreeFormObject