+++ kompendium-core/src/testFixtures/kotlin/io/bkbn/kompendium/core/fixtures/TestModels.kt
@@ -1,12 +1,26 @@
 package io.bkbn.kompendium.core.fixtures
 
-import java.util.UUID
-import io.bkbn.kompendium.annotations.KompendiumField
-import io.bkbn.kompendium.annotations.KompendiumParam
+import io.bkbn.kompendium.annotations.Field
+import io.bkbn.kompendium.annotations.FreeFormObject
+import io.bkbn.kompendium.annotations.Param
 import io.bkbn.kompendium.annotations.ParamType
 import io.bkbn.kompendium.annotations.UndeclaredField
+import io.bkbn.kompendium.annotations.constraint.Format
+import io.bkbn.kompendium.annotations.constraint.MaxItems
+import io.bkbn.kompendium.annotations.constraint.MaxLength
+import io.bkbn.kompendium.annotations.constraint.MaxProperties
+import io.bkbn.kompendium.annotations.constraint.Maximum
+import io.bkbn.kompendium.annotations.constraint.MinItems
+import io.bkbn.kompendium.annotations.constraint.MinLength
+import io.bkbn.kompendium.annotations.constraint.MinProperties
+import io.bkbn.kompendium.annotations.constraint.Minimum
+import io.bkbn.kompendium.annotations.constraint.MultipleOf
+import io.bkbn.kompendium.annotations.constraint.Pattern
+import io.bkbn.kompendium.annotations.constraint.UniqueItems
+import kotlinx.serialization.json.JsonElement
 import java.math.BigDecimal
 import java.math.BigInteger
+import java.util.UUID
 
 data class TestSimpleModel(val a: String, val b: Int)
 
@@ -27,8 +41,8 @@ data class TestSimpleWithEnumList(val a: Double, val b: List<SimpleEnum>)
 data class TestInvalidMap(val a: Map<Int, TestSimpleModel>)
 
 data class TestParams(
-  @KompendiumParam(ParamType.PATH) val a: String,
-  @KompendiumParam(ParamType.QUERY) val aa: Int
+  @Param(ParamType.PATH) val a: String,
+  @Param(ParamType.QUERY) val aa: Int
 )
 
 data class TestNested(val nesty: String)
@@ -36,7 +50,7 @@ data class TestNested(val nesty: String)
 data class TestWithUUID(val id: UUID)
 
 data class TestRequest(
-  @KompendiumField(name = "field_name")
+  @Field(name = "field_name")
   val fieldName: TestNested,
   val b: Double,
   val aaa: List<Long>
@@ -49,13 +63,101 @@ data class TestGeneric<T>(val messy: String, val potato: T)
 data class TestCreatedResponse(val id: Int, val c: String)
 
 data class TestFieldOverride(
-  @KompendiumField(name = "real_name", description = "A Field that is super important!")
+  @Field(name = "real_name", description = "A Field that is super important!")
   val b: Boolean
+)
+
+data class MinMaxInt(
+  @Minimum("5")
+  @Maximum("100")
+  val a: Int
+)
+
+data class MinMaxDouble(
+  @Minimum("5.5")
+  @Maximum("13.37")
+  val a: Double
+)
+
+data class ExclusiveMinMax(
+  @Minimum("5", true)
+  @Maximum("100", true)
+  val a: Int
+)
+
+data class FormattedString(
+  @Format("password")
+  @Param(ParamType.QUERY)
+  val a: String
+)
+
+data class MinMaxString(
+  @MinLength(42)
+  @MaxLength(1337)
+  val a: String
+)
+
+data class RegexString(
+  @Pattern("^\\d{3}-\\d{2}-\\d{4}\$")
+  val a: String
+)
+
+data class MinMaxArray(
+  @MinItems(1)
+  @MaxItems(10)
+  val a: List<String>
+)
+
+data class UniqueArray(
+  @UniqueItems
+  val a: List<Int>
+)
+
+data class MultipleOfInt(
+  @MultipleOf("5")
+  val a: Int
+)
+
+data class MultipleOfDouble(
+  @MultipleOf("2.5")
+  val a: Double
+)
+
+data class FreeFormData(
+  @FreeFormObject
+  val data: JsonElement
+)
+
+data class MinMaxFreeForm(
+  @FreeFormObject
+  @MinProperties(5)
+  @MaxProperties(10)
+  val data: JsonElement
+)
+
+
+data class RequiredParam(
+  @Param(ParamType.QUERY)
+  val a: String
+)
+
+data class DefaultParam(
+  @Param(ParamType.QUERY)
+  val b: String = "heyo"
+)
+
+data class DefaultField(
+  val a: String = "hi",
+  val b: Int
+)
+
+data class NullableField(
+  val a: String?
 )
 
 data class ComplexRequest(
   val org: String,
-  @KompendiumField("amazing_field")
+  @Field("amazing_field")
   val amazingField: String,
   val tables: List<NestedComplexItem>
 )
@@ -75,16 +177,16 @@ enum class SimpleEnum {
 }
 
 data class DefaultParameter(
-  @KompendiumParam(ParamType.QUERY) val a: Int = 100,
-  @KompendiumParam(ParamType.PATH) val b: String?,
-  @KompendiumParam(ParamType.PATH) val c: Boolean
+  @Param(ParamType.QUERY) val a: Int = 100,
+  @Param(ParamType.PATH) val b: String?,
+  @Param(ParamType.PATH) val c: Boolean
 )
 
 data class ExceptionResponse(val message: String)
 
 data class OptionalParams(
-  @KompendiumParam(ParamType.QUERY) val required: String,
-  @KompendiumParam(ParamType.QUERY) val notRequired: String?
+  @Param(ParamType.QUERY) val required: String,
+  @Param(ParamType.QUERY) val notRequired: String?
 )
 
 sealed class FlibbityGibbit
@@ -112,5 +214,5 @@ enum class Hehe {
 data class Mysterious(val nowYouSeeMe: String)
 
 data class HeaderNameTest(
-  @KompendiumParam(type = ParamType.HEADER) val `X-UserEmail`: String
+  @Param(type = ParamType.HEADER) val `X-UserEmail`: String
 )