+++ kompendium-locations/src/main/kotlin/io/bkbn/kompendium/locations/NotarizedLocation.kt
@@ -34,7 +34,7 @@ object NotarizedLocation {
   /**
    * Notarization for an HTTP GET request leveraging the Ktor [io.ktor.locations.Locations] plugin
    * @param TParam The class containing all parameter fields.
-   * Each field must be annotated with @[io.bkbn.kompendium.annotations.KompendiumParam].
+   * Each field must be annotated with @[io.bkbn.kompendium.annotations.Param].
    * Additionally, the class must be annotated with @[io.ktor.locations.Location].
    * @param TResp Class detailing the expected API response
    * @param info Route metadata
@@ -62,7 +62,7 @@ object NotarizedLocation {
   /**
    * Notarization for an HTTP POST request leveraging the Ktor [io.ktor.locations.Locations] plugin
    * @param TParam The class containing all parameter fields.
-   * Each field must be annotated with @[io.bkbn.kompendium.annotations.KompendiumParam]
+   * Each field must be annotated with @[io.bkbn.kompendium.annotations.Param]
    * Additionally, the class must be annotated with @[io.ktor.locations.Location].
    * @param TReq Class detailing the expected API request body
    * @param TResp Class detailing the expected API response
@@ -91,7 +91,7 @@ object NotarizedLocation {
   /**
    * Notarization for an HTTP Delete request leveraging the Ktor [io.ktor.locations.Locations] plugin
    * @param TParam The class containing all parameter fields.
-   * Each field must be annotated with @[io.bkbn.kompendium.annotations.KompendiumParam]
+   * Each field must be annotated with @[io.bkbn.kompendium.annotations.Param]
    * Additionally, the class must be annotated with @[io.ktor.locations.Location].
    * @param TReq Class detailing the expected API request body
    * @param TResp Class detailing the expected API response
@@ -120,7 +120,7 @@ object NotarizedLocation {
   /**
    * Notarization for an HTTP POST request leveraging the Ktor [io.ktor.locations.Locations] plugin
    * @param TParam The class containing all parameter fields.
-   * Each field must be annotated with @[io.bkbn.kompendium.annotations.KompendiumParam]
+   * Each field must be annotated with @[io.bkbn.kompendium.annotations.Param]
    * Additionally, the class must be annotated with @[io.ktor.locations.Location].
    * @param TResp Class detailing the expected API response
    * @param info Route metadata