+++ kompendium-locations/src/test/kotlin/io/bkbn/kompendium/locations/util/TestModels.kt
@@ -1,13 +1,13 @@
 package io.bkbn.kompendium.locations.util
 
-import io.bkbn.kompendium.annotations.KompendiumParam
+import io.bkbn.kompendium.annotations.Param
 import io.bkbn.kompendium.annotations.ParamType
 import io.ktor.locations.Location
 
 @Location("/test/{name}")
-data class SimpleLoc(@KompendiumParam(ParamType.PATH) val name: String) {
+data class SimpleLoc(@Param(ParamType.PATH) val name: String) {
   @Location("/nesty")
-  data class NestedLoc(@KompendiumParam(ParamType.QUERY) val isCool: Boolean, val parent: SimpleLoc)
+  data class NestedLoc(@Param(ParamType.QUERY) val isCool: Boolean, val parent: SimpleLoc)
 }
 
 data class SimpleResponse(val result: Boolean)