+++ kompendium-oas/src/main/kotlin/io/bkbn/kompendium/oas/path/PathOperation.kt
@@ -16,7 +16,7 @@ data class PathOperation(
   var parameters: List<Parameter>? = null,
   var requestBody: Request<*>? = null,
   // TODO How to enforce `default` requirement 🧐
-  var responses: Map<Int, Payload>? = null, // TODO Can be Response<*>?
+  var responses: Map<Int, Response<*>>? = null,
   var callbacks: Map<String, Payload>? = null, // todo what is this?
   var deprecated: Boolean = false,
   var security: List<Map<String, List<String>>>? = null,