+++ kompendium-oas/src/main/kotlin/io/bkbn/kompendium/oas/schema/DictionarySchema.kt
@@ -3,7 +3,8 @@ package io.bkbn.kompendium.oas.schema
 data class DictionarySchema(
   val additionalProperties: ComponentSchema,
   override val default: Any? = null,
-  override val description: String? = null
+  override val description: String? = null,
+  override val nullable: Boolean? = null
 ) : TypedSchema {
   override val type: String = "object"
 }