+++ kompendium-oas/src/main/kotlin/io/bkbn/kompendium/oas/schema/ObjectSchema.kt
@@ -3,7 +3,10 @@ package io.bkbn.kompendium.oas.schema
 data class ObjectSchema(
   val properties: Map<String, ComponentSchema>,
   override val default: Any? = null,
-  override val description: String? = null
+  override val description: String? = null,
+  override val nullable: Boolean? = null,
+  // constraints
+  val required: List<String>? = null
 ) : TypedSchema {
   override val type = "object"
 }