+++ kompendium-oas/src/main/kotlin/io/bkbn/kompendium/oas/schema/TypedSchema.kt
@@ -2,5 +2,6 @@ package io.bkbn.kompendium.oas.schema
 
 sealed interface TypedSchema : ComponentSchema {
   val type: String
+  val nullable: Boolean?
   override val default: Any?
 }